// frontend/src/app/models/korisnici.ts
export class korisnici{
    korIme: string = "";
    lozinka: string = "";
    ime: string = "";
    prezime: string = "";
    pol: string = "";
    adresa: string = "";
    kontaktTelefon: string = "";
    imejl: string = "";
    slika: any;
    kartica: string = "";
    tip: string = "";
    status: string = "";
}